Add auto trait for `DeserializeSlice` and `ParseHexStr`
- Modify tests accordingly
- Rearrange tests structure
