+++ dusk-bytes/src/serialize.rs
@@ -43,3 +43,7 @@ pub trait DeserializableSlice<const N: usize>: Serializable<N> {
         }
     }
 }
+
+// Implements automatically [`DeserializableSlice`] for any type that implements
+// [`Serializable`]
+impl<T, const N: usize> DeserializableSlice<N> for T where T: Serializable<N> {}
